break out is_error_count_healthy test
